+++ 05-ScenceManager/Fireball.cpp
@@ -19,7 +19,7 @@ CFireball::CFireball(D3DXVECTOR2 position, int nx)
 	this->SetAnimationSet(CAnimationSets::GetInstance()->Get(9));
 }
 
-void CFireball::Update(DWORD dt, vector<LPGAMEOBJECT>* coObjects)
+void CFireball::Update(ULONGLONG dt, vector<LPGAMEOBJECT>* coObjects)
 {
 	CGameObject::Update(dt);
