+++ 05-ScenceManager/GameObject.h
@@ -79,7 +79,7 @@ public:
 
 	int state;
 
-	DWORD dt; 
+	ULONGLONG dt; 
 	
 	LPANIMATION_SET animation_set;
 
@@ -115,7 +115,7 @@ public:
 	CGameObject();
 
 	virtual void GetBoundingBox(float &left, float &top, float &right, float &bottom) = 0;
-	virtual void Update(DWORD dt, vector<LPGAMEOBJECT> *coObjects = NULL);
+	virtual void Update(ULONGLONG dt, vector<LPGAMEOBJECT> *coObjects = NULL);
 	virtual void Render() = 0;
 	virtual void SetState(int state) { this->state = state; }
 	void ResetAni();
