+++ 05-ScenceManager/PlayScence.h
@@ -70,7 +70,7 @@ public:
 	CPlayScene(int id, LPCWSTR filePath);
 
 	virtual void Load();
-	virtual void Update(DWORD dt);
+	virtual void Update(ULONGLONG dt);
 	virtual void Render();
 	virtual void Unload();
