+++ 05-ScenceManager/Scence.h
@@ -16,7 +16,7 @@ public:
 	CKeyEventHandler * GetKeyEventHandler() { return key_handler; }
 	virtual void Load() = 0;
 	virtual void Unload() = 0;
-	virtual void Update(DWORD dt) = 0;
+	virtual void Update(ULONGLONG dt) = 0;
 	virtual void Render() = 0; 
 };
 typedef CScene * LPSCENE;
